split front, back led tests, put test separately
